move Slide of to seperate file of /Slide and make default export. Add export for SlideContent to /Slide index
